Add logging to trace run failure when best tap cannot be chosen for intermediate stop trip mode choice. Also added check to setHouseholds method in school location choice model to reset households if school location model results are not consistent with previous setHousehold method call.

+++ src/main/java/org/sandag/abm/ctramp/IntermediateStopChoiceModels.java
@@ -1183,7 +1183,7 @@ public class IntermediateStopChoiceModels
 	                    else {
 	                        
                         	//pick transit path from N-paths
-                        	float rn = (float)household.getHhRandom().nextDouble();
+                        	double rn = household.getHhRandom().nextDouble();
                         	int pathindex = logsumHelper.chooseTripPath(rn, segmentIkBestTapPairs[selectedIndex][accEgr], household.getDebugChoiceModels(), smcLogger);
                         	
                             stop.setBoardTap( (int)segmentIkBestTapPairs[selectedIndex][accEgr][pathindex][0] );
@@ -1642,7 +1642,7 @@ public class IntermediateStopChoiceModels
 	                    else {
 	                    	
 	                    	//pick transit path from N-paths
-                        	float rn = (float)household.getHhRandom().nextDouble();
+                        	double rn = household.getHhRandom().nextDouble();
                         	int pathindex = logsumHelper.chooseTripPath(rn, segmentIkBestTapPairs[selectedIndex][accEgr], household.getDebugChoiceModels(), smcLogger);
                         	
                             stop.setBoardTap( (int)segmentIkBestTapPairs[selectedIndex][accEgr][pathindex][0] );
